requirement.txt
argon2==0.1.10
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
ecdsa==0.14.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
google==3.0.0
grpcio==1.32.0
grpcio-tools==1.32.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libsignal==0.0.4
MarkupSafe==1.1.1
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
python-axolotl-curve25519==0.4.1.post2
python-jose==3.2.0
python-keycloak==0.22.0
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
